requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
lightgbm
pyyaml
joblib
shap
python-dateutil
